long long

+++ others/measure_time_elapsed.cpp
@@ -1,7 +1,7 @@
 // To calculate the time taken by a code to execute
 #include <sys/time.h>
 
-int64_t getTimeInMicroseconds() {
+long long getTimeInMicroseconds() {
     struct timeval start;
     gettimeofday(&start, NULL);
     return start.tv_sec * 1000000 + start.tv_usec;
@@ -11,7 +11,7 @@ int64_t getTimeInMicroseconds() {
 
 int main() {
     // write code
-    int64_t starttime = getTimeInMicroseconds();
+    long long starttime = getTimeInMicroseconds();
     // sample(args) function run
     // Any other functions (if present) run
     std:cout << getTimeInMicroseconds() - starttime;
